pyproject.toml
[tool.poetry]
name = "domain_finder"
version = "0.1.0"
description = "Coding assignment for webloom"
authors = ["Randeeep Singh"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.0.5"
djangorestframework = "^3.13.1"
django-filter = "^21.1"
requests = "^2.27.1"
asyncwhois = "1.0.0"

[tool.poetry.dev-dependencies]
pylint = "^2.13.9"
black = "^22.3.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
